requirements.txt
Flask==1.0.2
termcolor==1.1.0
Werkzeug==0.14.1
PyYAML==3.12
python-dateutil
gunicorn==19.8.1
pytz
requests
urllib3
pytest==3.6.0
redis==2.10.6
rq==0.10.0
Flask-And-Redis==0.7
celery==4.1.1

numpy
scipy
matplotlib
psychrochart
homeassistant
